+++ src/nextjs/utils/withMiddlewareAuthRequired.ts
@@ -3,7 +3,10 @@ import { NextResponse } from 'next/dist/server/web/spec-extension/response';
 import { NextMiddleware } from 'next/server';
 import { User, ApiError, createClient } from '@supabase/supabase-js';
 import { CookieOptions } from '../types';
-import { COOKIE_OPTIONS } from '../../shared/utils/constants';
+import {
+  COOKIE_OPTIONS,
+  TOKEN_REFRESH_MARGIN
+} from '../../shared/utils/constants';
 import { jwtDecoder } from '../../shared/utils/jwt';
 import { setCookies } from '../../shared/utils/cookies';
 import {
@@ -11,7 +14,7 @@ import {
   NextResponseAdapter
 } from '../../shared/adapters/NextMiddlewareAdapter';
 
-export type WithMiddlewareAuthRequired = (options?: {
+export interface WithMiddlewareAuthRequiredOptions {
   /**
    * Path relative to the site root to redirect an
    * unauthenticated visitor.
@@ -21,10 +24,14 @@ export type WithMiddlewareAuthRequired = (options?: {
    */
   redirectTo?: string;
   cookieOptions?: CookieOptions;
-}) => NextMiddleware;
+  tokenRefreshMargin?: number;
+}
+export type WithMiddlewareAuthRequired = (
+  options?: WithMiddlewareAuthRequiredOptions
+) => NextMiddleware;
 
 export const withMiddlewareAuthRequired: WithMiddlewareAuthRequired =
-  (options: { redirectTo?: string; cookieOptions?: CookieOptions } = {}) =>
+  (options: WithMiddlewareAuthRequiredOptions = {}) =>
   async (req) => {
     try {
       if (
@@ -44,6 +51,8 @@ export const withMiddlewareAuthRequired: WithMiddlewareAuthRequired =
         { fetch }
       );
       const cookieOptions = { ...COOKIE_OPTIONS, ...options.cookieOptions };
+      const tokenRefreshMargin =
+        options.tokenRefreshMargin ?? TOKEN_REFRESH_MARGIN;
       const access_token = req.cookies[`${cookieOptions.name!}-access-token`];
       const refresh_token = req.cookies[`${cookieOptions.name!}-refresh-token`];
 
@@ -62,7 +71,7 @@ export const withMiddlewareAuthRequired: WithMiddlewareAuthRequired =
           throw new Error('Not able to parse JWT payload!');
         }
         const timeNow = Math.round(Date.now() / 1000);
-        if (jwtUser.exp < timeNow) {
+        if (jwtUser.exp < timeNow + tokenRefreshMargin) {
           if (!refresh_token) {
             throw new Error('No refresh_token cookie found!');
           }